finish permissions

+++ e-commerce/backend/src/resolvers/Mutation.js
@@ -47,9 +47,11 @@ const mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id, title}`)
+    const item = await ctx.db.query.item({ where }, `{id, title, user { id }}`)
     // check if they own that item o have the permissions
-
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+    if (!ownsItem && !hasPermissions) throw new Error("You don't have permission to do that!")
     //delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
